pyproject.toml
[tool.poetry]
name = "text-statistical-analysis"
version = "0.1.0"
description = ""
authors = ["Sergey Yaksanov <iaksanov2044@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django = "^5.1.4"
django-cors-headers = "^4.6.0"
python-decouple = "^3.8"
django-stubs = "^5.1.1"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
